修改：人才培养：学习资源

+++ person-training/src/main/java/com/itts/personTraining/service/kssj/impl/KssjServiceImpl.java
@@ -342,7 +342,7 @@ public class KssjServiceImpl extends ServiceImpl<KssjMapper, Kssj> implements Ks
     private void random(Long id,Sjtxndpz sjtxndpz,String sjtxnd,String tkzyType,List<Long> kcIdList){
         List<Tkzy> by = getBy(sjtxnd, tkzyType, kcIdList);
         //随机获取题目
-        if(by.size() < sjtxndpz.getTs()){
+        if(by.size() < sjtxndpz.getTs().intValue()){
             Kssj kssj = kssjMapper.selectOne(new QueryWrapper<Kssj>().eq("id", id).eq("sfsc", false));
             kssj.setSfsc(true);
             kssjMapper.updateById(kssj);
